refactor(treeView): clean up files
