--- client/src/Components/Tree/Tree.module.css
@@ -1,6 +1,0 @@
-.tree {
-    display: flex;
-    flex-direction: column;
-    flex-wrap: wrap;
-    white-space: nowrap;
-}
