+++ client/src/Components/TreeView/TreeView.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Tree from "../Tree/Tree";
-import "./TreeView.module.css";
 import PythonPackage from "../../model/python/PythonPackage";
+import {isEmptyList} from "../../util/listOperations";
+import ModuleNode from "./ModuleNode";
 import TreeViewCSS from "./TreeView.module.css";
 
 interface TreeViewProps {
@@ -11,7 +11,14 @@ interface TreeViewProps {
 export default function TreeView(props: TreeViewProps): JSX.Element {
     return (
         <div className={TreeViewCSS.treeView}>
-            <Tree pythonPackage={props.pythonPackage}/>
+            {[...props.pythonPackage.modules]
+                .sort((a, b) => a.name.localeCompare(b.name))
+                .filter(module => !isEmptyList(module.classes) || !isEmptyList(module.functions))
+                .map(module => (
+                    <ModuleNode key={module.name}
+                                pythonModule={module}/>
+                ))
+            }
         </div>
     );
 }
